version two

+++ Total_code/Movement.h
@@ -1,20 +1,19 @@
 #ifndef MOVEMENT_FUNCTIONS_H
 #define MOVEMENT_FUNCTIONS_H
 
-//#include <Wire.h>
-//#include <Adafruit_MPU6050.h>
-//#include <Adafruit_Sensor.h>
 #include <math.h>
+#include "encoder.h"
+#include "gyroscope.h"
 
-//motors macros
+// Motor control pin definitions
 #define MOTOR_RIGHT_PIN1 5
 #define MOTOR_RIGHT_PIN2 18
 #define MOTOR_LEFT_PIN1 4
 #define MOTOR_LEFT_PIN2 0
 
-void moveForward(int distance);
+// Function declarations
+void moveForward(float distance);
 void turnLeft();
 void turnRight();
 
-
-#endif
+#endif // MOVEMENT_FUNCTIONS_H
